+++ WargameEngine/WargameEngine/WargameEngine/NetSocket.h
@@ -8,8 +8,8 @@ public:
 	virtual void InitHost(unsigned short port = 0) override;
 	virtual void InitClient(const char * ip, unsigned short port = 0) override;
 	virtual void InitFromAnotherSocket(unsigned int socket, void* sockAddr) override;
-	virtual bool SendData(const char * data, int len) const override;
-	virtual int RecieveData(char* data, int maxLength) override;
+	virtual bool SendData(const char * data, size_t len) const override;
+	virtual int RecieveData(char* data, size_t maxLength) override;
 	std::string GetIP() const;
 	unsigned short GetPort() const;
 	~CNetSocket();
